pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reporting-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.85.0",
    "python-dotenv>=1.0.0",
    "openai>=2.29.0",
    "qdrant-client>=1.17.1",
    "pyodbc>=5.3.0",
]

[project.scripts]
reporting-agent = "agent.loop:main"

[tool.hatch.build.targets.wheel]
packages = ["agent", "services"]
